steps generation change, update balance

+++ back/src/shd/maths/steps.rs
@@ -18,10 +18,10 @@ pub struct OnChainLiquidity {
  * Sum the total liquidity of a pair of tokens.
  * @dev components Every similar components (= matching a pair)
  */
-pub fn deepth(components: Vec<SrzProtocolComponent>, targets: Vec<SrzToken>, data: HashMap<String, HashMap<String, u128>>) -> HashMap<String, u128> {
+pub fn deepth(components: Vec<SrzProtocolComponent>, targets: Vec<SrzToken>, data: HashMap<String, HashMap<String, f64>>) -> HashMap<String, f64> {
     let mut cumulated = HashMap::new();
     targets.iter().for_each(|t| {
-        cumulated.insert(t.clone().address.to_lowercase(), 0);
+        cumulated.insert(t.clone().address.to_lowercase(), 0f64);
     });
     // Every component containing 'tokens'
     for cp in components.clone().iter() {
